+++ src/solver/optim-model-filler/include/antares/solver/optim-model-filler/LinearExpression.h
@@ -19,7 +19,7 @@
  * along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
  */
 #pragma once
-#include <span>
+
 #include <utility>
 #include <variant>
 #include <vector>
@@ -55,45 +55,4 @@ private:
     double constant_ = 0.;
 };
 
-class TimeDependentLinearExpression final
-{
-public:
-    explicit TimeDependentLinearExpression(std::size_t nbTimesteps);
-
-    explicit TimeDependentLinearExpression(const std::span<const double>& values);
-
-    explicit TimeDependentLinearExpression(LinearExpression&& expr);
-    void expandTo(std::size_t nbTimesteps);
-
-    std::vector<double> constant() const;
-
-    void mergeDuplicateCoefficients();
-
-    std::size_t size() const;
-
-    LinearExpression* begin();
-    LinearExpression* end();
-
-    const LinearExpression* begin() const;
-    const LinearExpression* end() const;
-
-    LinearExpression& operator[](std::size_t idx);
-
-    const LinearExpression& operator[](std::size_t idx) const;
-
-    TimeDependentLinearExpression& operator+=(const TimeDependentLinearExpression& other);
-
-    TimeDependentLinearExpression& operator-=(const TimeDependentLinearExpression& other);
-
-    void rotate(int shift);
-
-    TimeDependentLinearExpression& operator*=(double factor);
-
-    TimeDependentLinearExpression& operator*=(const TimeDependentLinearExpression& other);
-
-    TimeDependentLinearExpression operator-() const;
-
-private:
-    std::variant<LinearExpression, std::vector<LinearExpression>> v_;
-};
 } // namespace Antares::Optimization
