+++ src/solver/optim-model-filler/include/antares/solver/optim-model-filler/TimeDependentLinearExpression.h
@@ -1,0 +1,71 @@
+/*
+ * Copyright 2007-2024, RTE (https://www.rte-france.com)
+ * See AUTHORS.txt
+ * SPDX-License-Identifier: MPL-2.0
+ * This file is part of Antares-Simulator,
+ * Adequacy and Performance assessment for interconnected energy networks.
+ *
+ * Antares_Simulator is free software: you can redistribute it and/or modify
+ * it under the terms of the Mozilla Public Licence 2.0 as published by
+ * the Mozilla Foundation, either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * Antares_Simulator is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * Mozilla Public Licence 2.0 for more details.
+ *
+ * You should have received a copy of the Mozilla Public Licence 2.0
+ * along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
+ */
+#pragma once
+#include <span>
+
+#include "LinearExpression.h"
+
+namespace Antares::Optimization
+{
+class TimeDependentLinearExpression final
+{
+public:
+    explicit TimeDependentLinearExpression(std::size_t nbTimesteps);
+
+    explicit TimeDependentLinearExpression(const std::span<const double>& values);
+
+    explicit TimeDependentLinearExpression(LinearExpression&& expr);
+
+    std::vector<double> constant() const;
+
+    void mergeDuplicateCoefficients();
+
+    std::size_t size() const;
+
+    LinearExpression* begin();
+    LinearExpression* end();
+
+    const LinearExpression* begin() const;
+    const LinearExpression* end() const;
+
+    LinearExpression& operator[](std::size_t idx);
+
+    const LinearExpression& operator[](std::size_t idx) const;
+
+    TimeDependentLinearExpression& operator+=(const TimeDependentLinearExpression& other);
+
+    TimeDependentLinearExpression& operator-=(const TimeDependentLinearExpression& other);
+
+    void rotate(int shift);
+
+    TimeDependentLinearExpression& operator*=(double factor);
+
+    TimeDependentLinearExpression& operator*=(const TimeDependentLinearExpression& other);
+
+    TimeDependentLinearExpression operator-() const;
+
+private:
+    void expandTo(std::size_t nbTimesteps);
+
+    std::variant<LinearExpression, std::vector<LinearExpression>> v_;
+};
+
+} // namespace Antares::Optimization
